pyproject.toml
[project]
name = "icair-certificate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cloudinary>=1.44.1",
    "flask>=3.1.2",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.4",
    "requests>=2.32.5",
]
